Split out the insert stmt values
